factored out mutation status data retrieving function into MutationStatusColumnFormatter
